+++ lib/src/models/transactions/types/asset_config_transaction.dart
@@ -2,7 +2,6 @@ import 'dart:typed_data';
 
 import 'package:algorand_dart/src/models/models.dart';
 import 'package:algorand_dart/src/models/transactions/builders/transaction_builders.dart';
-import 'package:algorand_dart/src/utils/serializers/bigint_serializer.dart';
 import 'package:algorand_dart/src/utils/serializers/serializers.dart';
 import 'package:algorand_dart/src/utils/transformers/address_transformer.dart';
 import 'package:json_annotation/json_annotation.dart';
@@ -112,7 +111,7 @@ class AssetConfigTransaction extends RawTransaction {
     this.freezeAddress,
     this.clawbackAddress,
     this.destroy = false,
-    int? fee,
+    BigInt? fee,
     int? firstValid,
     Uint8List? genesisHash,
     int? lastValid,
