+++ lib/src/models/transactions/types/asset_transfer_transaction.dart
@@ -1,7 +1,6 @@
 import 'dart:typed_data';
 
 import 'package:algorand_dart/algorand_dart.dart';
-import 'package:algorand_dart/src/utils/serializers/bigint_serializer.dart';
 import 'package:algorand_dart/src/utils/transformers/address_transformer.dart';
 import 'package:json_annotation/json_annotation.dart';
 
@@ -48,7 +47,7 @@ class AssetTransferTransaction extends RawTransaction {
     this.assetSender,
     this.receiver,
     this.closeTo,
-    int? fee,
+    BigInt? fee,
     int? firstValid,
     Uint8List? genesisHash,
     int? lastValid,
